PR Responses and other cleanups:
 * Share _temp_database_url and _database_url between load.py and extract.py using SqlAlchemyMixin
 * Remove useless local_ids initialization in _execute_step
 * Make get_batch_iterator into an officially reusable util
 * Move MockBulkApiDmlOperation and MockBulkAPI to util file and rename to Fake*
 * Rename FakeSF.__getattr__ because it isn't like simple-salesfore.__getattr__
